+++ envisage/resource/http_resource_protocol.py
@@ -37,9 +37,7 @@ class HTTPResourceProtocol(HasTraits):
             f = urlopen("http://" + address)
 
         except HTTPError as exc:
-            # On Python 3.8 and 3.9, close() fails if the HTTPError has no fp.
-            if exc.fp is not None:
-                exc.close()
+            exc.close()
             raise NoSuchResourceError("http://" + address)
 
         return f